First Design for the Web site
In this file you will see the first base design for the whole web site, including three buttons, 2 articles and an image as a background. The propuse of this commit is to mantain a general first design that would be adapted for the rest of the html files that are going to be added later.