Fix connection problems
After looking for the connection between "Dormir" and "Control", I found that they were not connected, so I fix that problem. I also add to the web sites that had options to choose for the what if of the story an aside in order to separate them for a correct semantic use.